requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-dotenv==1.0.0
pydantic>=1.10.9
PyJWT==2.6.0
motor==3.2.0
aioredis==2.0.1
pymongo==4.4.0
redis>=4.5.4
requests==2.31.0
CMRESHandler==1.0.0
APScheduler==3.11.0
python-multipart==0.0.20
pytest==8.3.5
mongomock==4.3.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.40
starlette==0.26.1
slowapi==0.1.9
httpx==0.28.1
